refactor: consolidate models and move checksum type
- Consolidate asset and release models into core/github/models.py, rename models/selector.py to asset_selector.py, and remove the old models package __init__.
- Move ChecksumFileInfo into my_unicorn.types and update verification modules to import the type from my_unicorn.types.
