+++ src/my_unicorn/core/github/__init__.py
@@ -1,16 +1,12 @@
 """GitHub infrastructure - API client and release fetching."""
 
+from my_unicorn.core.github.asset_selector import AssetSelector
 from my_unicorn.core.github.client import ReleaseAPIClient
 from my_unicorn.core.github.config_service import (
     GitHubConfig,
     get_github_config,
 )
-from my_unicorn.core.github.models import (
-    Asset,
-    AssetSelector,
-    ChecksumFileInfo,
-    Release,
-)
+from my_unicorn.core.github.models import Asset, Release
 from my_unicorn.core.github.operations import (
     extract_github_config,
     parse_github_url,
@@ -20,7 +16,6 @@ from my_unicorn.core.github.release_fetcher import GitHubClient, ReleaseFetcher
 __all__ = [
     "Asset",
     "AssetSelector",
-    "ChecksumFileInfo",
     "GitHubClient",
     "GitHubConfig",
     "Release",
